contentApp/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
content==0.0.1
cycler==0.10.0
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.1
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
prettytable==1.0.1
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
scipy==1.5.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
wcwidth==0.2.5
